Move FFI wrappers to separate mod and remove Go example
