+++ README.md
@@ -17,6 +17,7 @@ This repo is composed of multiple crates:
 * temporal-sdk-core-api `./core-api` - Defines the API surface exposed by Core
 * temporal-sdk-core `./core` - The Core implementation
 * temporal-sdk `./sdk` - A (currently prototype) Rust SDK built on top of Core. Used for testing.
+* temporal-sdk-core-bridge-ffi `./bridge-ffi` - C API wrapper for Core
 * rustfsm `./fsm` - Implements a procedural macro used by core for defining state machines
     (contains subcrates). It is temporal agnostic.
