--- bridge-ffi/example/goffi/.gitignore
@@ -1,5 +1,0 @@
-# Don't commit example generated protos
-/corepb
-
-# Don't commit example static library
-lib/*/libtemporal_*
