+++ bridge-ffi/src/lib.rs
@@ -7,11 +7,14 @@
     // functions as unsafe. Even the example in clippy's docs at
     // https://rust-lang.github.io/rust-clippy/master/index.html#not_unsafe_ptr_arg_deref
     // cause a rustc warning for unnecessary inner-unsafe when marked on fn.
+    // This check only applies to "pub" functions which are all exposed via C
+    // API.
     clippy::not_unsafe_ptr_arg_deref,
 )]
 
+mod wrappers;
+
 use prost::Message;
-use std::str::FromStr;
 use temporal_sdk_core_protos::coresdk::bridge;
 
 extern crate libc;
@@ -52,8 +55,8 @@ impl tmprl_bytes_t {
     }
 }
 
-// We only impl this because they are required by lazy_static and raw pointers
-// are not usually safe for send
+/// Required because these instances are used by lazy_static and raw pointers
+/// are not usually safe for send/sync.
 unsafe impl Send for tmprl_bytes_t {}
 unsafe impl Sync for tmprl_bytes_t {}
 
@@ -94,8 +97,8 @@ pub extern "C" fn tmprl_bytes_free(core: *mut tmprl_core_t, bytes: *const tmprl_
     }
 }
 
-// Used for maintaining pointer to user data across threads. See
-// https://doc.rust-lang.org/nomicon/send-and-sync.html
+/// Used for maintaining pointer to user data across threads. See
+/// https://doc.rust-lang.org/nomicon/send-and-sync.html.
 struct UserDataHandle(*mut libc::c_void);
 unsafe impl Send for UserDataHandle {}
 unsafe impl Sync for UserDataHandle {}
@@ -232,7 +235,12 @@ pub extern "C" fn tmprl_core_init(
     };
     let user_data = UserDataHandle(user_data);
     runtime.tokio_runtime.spawn(async move {
-        match tmprl_core_t::new(runtime.tokio_runtime.clone(), CoreInitOptions(req)).await {
+        match tmprl_core_t::new(
+            runtime.tokio_runtime.clone(),
+            wrappers::CoreInitOptions(req),
+        )
+        .await
+        {
             Ok(core) => unsafe {
                 callback(
                     user_data.into(),
@@ -262,11 +270,11 @@ pub extern "C" fn tmprl_core_init(
 /// protobuf message. The callback is invoked on completion with a
 /// ShutdownResponse protobuf message.
 #[no_mangle]
-#[allow(unused_variables)] // We intentionally ignore the request
 pub extern "C" fn tmprl_core_shutdown(
     core: *mut tmprl_core_t,
+    #[allow(unused_variables)] // We intentionally ignore the request
     req_proto: *const u8,
-    req_proto_len: libc::size_t,
+    #[allow(unused_variables)] req_proto_len: libc::size_t,
     user_data: *mut libc::c_void,
     callback: tmprl_callback,
 ) {
@@ -314,7 +322,7 @@ pub extern "C" fn tmprl_register_worker(
         };
     let user_data = UserDataHandle(user_data);
     core.tokio_runtime.clone().spawn(async move {
-        match core.register_worker(WorkerConfig(req)).await {
+        match core.register_worker(wrappers::WorkerConfig(req)).await {
             Ok(()) => unsafe {
                 callback(user_data.into(), &*DEFAULT_REGISTER_WORKER_RESPONSE_BYTES);
             },
@@ -332,11 +340,11 @@ pub extern "C" fn tmprl_register_worker(
 /// ShutdownWorkerRequest protobuf message. The callback is invoked on
 /// completion with a ShutdownWorkerResponse protobuf message.
 #[no_mangle]
-#[allow(unused_variables)]
 pub extern "C" fn tmprl_shutdown_worker(
     core: *mut tmprl_core_t,
+    #[allow(unused_variables)] // We intentionally ignore the request
     req_proto: *const u8,
-    req_proto_len: libc::size_t,
+    #[allow(unused_variables)] req_proto_len: libc::size_t,
     user_data: *mut libc::c_void,
     callback: tmprl_callback,
 ) {
@@ -640,11 +648,11 @@ pub extern "C" fn tmprl_request_workflow_eviction(
 /// FetchBufferedLogsRequest protobuf message. The callback is invoked
 /// on completion with a FetchBufferedLogsResponse protobuf message.
 #[no_mangle]
-#[allow(unused_variables)] // We intentionally ignore the request
 pub extern "C" fn tmprl_fetch_buffered_logs(
     core: *mut tmprl_core_t,
+    #[allow(unused_variables)] // We intentionally ignore the request
     req_proto: *const u8,
-    req_proto_len: libc::size_t,
+    #[allow(unused_variables)] req_proto_len: libc::size_t,
     user_data: *mut libc::c_void,
     callback: tmprl_callback,
 ) {
@@ -661,7 +669,7 @@ pub extern "C" fn tmprl_fetch_buffered_logs(
 impl tmprl_core_t {
     async fn new(
         tokio_runtime: std::sync::Arc<tokio::runtime::Runtime>,
-        opts: CoreInitOptions,
+        opts: wrappers::CoreInitOptions,
     ) -> Result<tmprl_core_t, String> {
         Ok(tmprl_core_t {
             tokio_runtime,
@@ -679,7 +687,7 @@ impl tmprl_core_t {
 
     async fn register_worker(
         &self,
-        config: WorkerConfig,
+        config: wrappers::WorkerConfig,
     ) -> Result<(), bridge::register_worker_response::Error> {
         let config =
             config
@@ -821,196 +829,5 @@ impl tmprl_core_t {
     {
         P::decode(unsafe { std::slice::from_raw_parts(bytes, bytes_len) })
             .map_err(|err| format!("failed decoding proto: {}", err))
-    }
-}
-
-// Present for try-from only
-struct CoreInitOptions(bridge::InitRequest);
-
-impl TryFrom<CoreInitOptions> for temporal_sdk_core::CoreInitOptions {
-    type Error = String;
-
-    fn try_from(CoreInitOptions(req): CoreInitOptions) -> Result<Self, Self::Error> {
-        let mut core_opts = temporal_sdk_core::CoreInitOptionsBuilder::default();
-        if let Some(req_gateway_opts) = req.gateway_options {
-            let mut gateway_opts = temporal_sdk_core::ServerGatewayOptionsBuilder::default();
-            if !req_gateway_opts.target_url.is_empty() {
-                gateway_opts.target_url(
-                    temporal_sdk_core::Url::parse(&req_gateway_opts.target_url)
-                        .map_err(|err| format!("invalid target URL: {}", err))?,
-                );
-            }
-            if !req_gateway_opts.namespace.is_empty() {
-                gateway_opts.namespace(req_gateway_opts.namespace);
-            }
-            if !req_gateway_opts.client_name.is_empty() {
-                gateway_opts.client_name(req_gateway_opts.client_name);
-            }
-            if !req_gateway_opts.client_version.is_empty() {
-                gateway_opts.client_version(req_gateway_opts.client_version);
-            }
-            if !req_gateway_opts.static_headers.is_empty() {
-                gateway_opts.static_headers(req_gateway_opts.static_headers);
-            }
-            if !req_gateway_opts.identity.is_empty() {
-                gateway_opts.identity(req_gateway_opts.identity);
-            }
-            if !req_gateway_opts.worker_binary_id.is_empty() {
-                gateway_opts.worker_binary_id(req_gateway_opts.worker_binary_id);
-            }
-            if let Some(req_tls_config) = req_gateway_opts.tls_config {
-                let mut tls_config = temporal_sdk_core::TlsConfig::default();
-                if !req_tls_config.server_root_ca_cert.is_empty() {
-                    tls_config.server_root_ca_cert = Some(req_tls_config.server_root_ca_cert);
-                }
-                if !req_tls_config.domain.is_empty() {
-                    tls_config.domain = Some(req_tls_config.domain);
-                }
-                if !req_tls_config.client_cert.is_empty()
-                    || !req_tls_config.client_private_key.is_empty()
-                {
-                    tls_config.client_tls_config = Some(temporal_sdk_core::ClientTlsConfig {
-                        client_cert: req_tls_config.client_cert,
-                        client_private_key: req_tls_config.client_private_key,
-                    })
-                }
-                gateway_opts.tls_cfg(tls_config);
-            }
-            if let Some(req_retry_config) = req_gateway_opts.retry_config {
-                let mut retry_config = temporal_sdk_core::RetryConfig::default();
-                if let Some(v) = req_retry_config.initial_interval {
-                    retry_config.initial_interval =
-                        v.try_into().map_err(|_| "invalid initial interval")?;
-                }
-                if let Some(v) = req_retry_config.randomization_factor {
-                    retry_config.randomization_factor = v;
-                }
-                if let Some(v) = req_retry_config.multiplier {
-                    retry_config.multiplier = v;
-                }
-                if let Some(v) = req_retry_config.max_interval {
-                    retry_config.max_interval = v.try_into().map_err(|_| "invalid max interval")?;
-                }
-                if let Some(v) = req_retry_config.max_elapsed_time {
-                    retry_config.max_elapsed_time =
-                        Some(v.try_into().map_err(|_| "invalid max elapsed time")?);
-                }
-                if let Some(v) = req_retry_config.max_retries {
-                    retry_config.max_retries = v as usize;
-                }
-                gateway_opts.retry_config(retry_config);
-            }
-            core_opts.gateway_opts(
-                gateway_opts
-                    .build()
-                    .map_err(|err| format!("invalid gateway options: {}", err))?,
-            );
-        }
-        if let Some(req_telemetry_opts) = req.telemetry_options {
-            let mut telemetry_opts = temporal_sdk_core::TelemetryOptionsBuilder::default();
-            if !req_telemetry_opts.otel_collector_url.is_empty() {
-                telemetry_opts.otel_collector_url(
-                    temporal_sdk_core::Url::parse(&req_telemetry_opts.otel_collector_url)
-                        .map_err(|err| format!("invalid OpenTelemetry collector URL: {}", err))?,
-                );
-            }
-            if !req_telemetry_opts.tracing_filter.is_empty() {
-                telemetry_opts.tracing_filter(req_telemetry_opts.tracing_filter.clone());
-            }
-            match req_telemetry_opts.log_forwarding_level() {
-                bridge::LogLevel::Unspecified => {}
-                bridge::LogLevel::Off => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Off);
-                }
-                bridge::LogLevel::Error => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Error);
-                }
-                bridge::LogLevel::Warn => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Warn);
-                }
-                bridge::LogLevel::Info => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Info);
-                }
-                bridge::LogLevel::Debug => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Debug);
-                }
-                bridge::LogLevel::Trace => {
-                    telemetry_opts.log_forwarding_level(log::LevelFilter::Trace);
-                }
-            }
-            if !req_telemetry_opts.prometheus_export_bind_address.is_empty() {
-                telemetry_opts.prometheus_export_bind_address(
-                    std::net::SocketAddr::from_str(
-                        &req_telemetry_opts.prometheus_export_bind_address,
-                    )
-                    .map_err(|err| format!("invalid Prometheus address: {}", err))?,
-                );
-            }
-            core_opts.telemetry_opts(
-                telemetry_opts
-                    .build()
-                    .map_err(|err| format!("invalid telemetry options: {}", err))?,
-            );
-        }
-        core_opts
-            .build()
-            .map_err(|err| format!("invalid options: {}", err))
-    }
-}
-
-// Present for try-from only
-struct WorkerConfig(bridge::RegisterWorkerRequest);
-
-impl TryFrom<WorkerConfig> for temporal_sdk_core_api::worker::WorkerConfig {
-    type Error = String;
-
-    fn try_from(WorkerConfig(req): WorkerConfig) -> Result<Self, Self::Error> {
-        let mut config = temporal_sdk_core_api::worker::WorkerConfigBuilder::default();
-        if !req.task_queue.is_empty() {
-            config.task_queue(req.task_queue);
-        }
-        if let Some(v) = req.max_cached_workflows {
-            config.max_cached_workflows(v as usize);
-        }
-        if let Some(v) = req.max_outstanding_workflow_tasks {
-            config.max_outstanding_workflow_tasks(v as usize);
-        }
-        if let Some(v) = req.max_outstanding_activities {
-            config.max_outstanding_activities(v as usize);
-        }
-        if let Some(v) = req.max_outstanding_local_activities {
-            config.max_outstanding_local_activities(v as usize);
-        }
-        if let Some(v) = req.max_concurrent_wft_polls {
-            config.max_concurrent_wft_polls(v as usize);
-        }
-        if let Some(v) = req.nonsticky_to_sticky_poll_ratio {
-            config.nonsticky_to_sticky_poll_ratio(v);
-        }
-        if let Some(v) = req.max_concurrent_at_polls {
-            config.max_concurrent_at_polls(v as usize);
-        }
-        config.no_remote_activities(req.no_remote_activities);
-        if let Some(v) = req.sticky_queue_schedule_to_start_timeout {
-            let v: std::time::Duration = v
-                .try_into()
-                .map_err(|_| "invalid sticky queue schedule to start timeout".to_string())?;
-            config.sticky_queue_schedule_to_start_timeout(v);
-        }
-        if let Some(v) = req.max_heartbeat_throttle_interval {
-            let v: std::time::Duration = v
-                .try_into()
-                .map_err(|_| "invalid max heartbeat throttle interval".to_string())?;
-            config.max_heartbeat_throttle_interval(v);
-        }
-        if let Some(v) = req.default_heartbeat_throttle_interval {
-            let v: std::time::Duration = v
-                .try_into()
-                .map_err(|_| "invalid default heartbeat throttle interval".to_string())?;
-            config.default_heartbeat_throttle_interval(v);
-        }
-        config
-            .build()
-            .map_err(|err| format!("invalid request: {}", err))
     }
 }
